requirements.txt
backcall==0.2.0
colorama==0.4.6
debugpy==1.6.6
decorator==5.1.1
entrypoints==0.4
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.2
jupyter_client==7.4.9
jupyter_core==4.12.0
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
numpy==1.21.6
packaging==23.0
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.37
protobuf==3.20.3
psutil==5.9.4
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
six==1.16.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
wcwidth==0.2.6
